requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.75
botocore==1.34.84
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jmespath==1.0.1
Js2Py==0.74
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
pyjsparser==2.7.1
PyJWT==2.8.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.11.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
